Fix: 뷰모델 하나로 합치기
